+++ _examples/core/cmd_test.go
@@ -1,0 +1,11 @@
+package core
+
+import (
+	"fmt"
+	"github.com/xingcxb/goKit/core/runTimeKit"
+	"testing"
+)
+
+func TestShell(t *testing.T) {
+	fmt.Println(runTimeKit.ExecuteCmd("curl", "cip.cc/172.42.102.198"))
+}
