+++ core/strKit/idCardKit.go
@@ -1,0 +1,2 @@
+// Package strKit 身份证
+package strKit
